Refactor JS modules to be consumed with Mojular
