--- assets/scripts/modules/moj.skip-to-content.js
@@ -1,18 +1,0 @@
-(function () {
-  'use strict';
-
-  moj.Modules.SkipToContent = {
-    init: function() {
-      $('body').on('click', '.skip-link', function(e) {
-        e.preventDefault();
-
-        $($(e.target).attr('href'))
-          .attr('tabindex', -1)
-          .focus()
-          .blur(function() {
-            $(this).removeAttr('tabindex');
-          });
-      });
-    }
-  };
-}());
