+++ assets/scripts/modules/tabs.js
