+++ package.json
@@ -21,7 +21,11 @@
   },
   "homepage": "https://github.com/mojular/moj#readme",
   "dependencies": {
+    "mojular": "mojular/mojular#webpack",
     "mojular-govuk-elements": "mojular/govuk-elements#dev"
+  },
+  "devDependencies": {
+    "json-loader": "^0.5.3"
   },
   "paths": {
     "sass": [
